Send entries in batches

+++ contracts/interfaces/IRewardEscrowV2.sol
@@ -89,7 +89,7 @@ interface IRewardEscrowV2 {
     ) external;
 
     // Return amount of SNX transfered to SynthetixBridgeToOptimism deposit contract
-    function burnForMigration(address account, uint[] calldata entryIDs)
+    function burnForMigration(address account, uint256[] calldata entryIDs)
         external
         returns (uint256 escrowedAccountBalance, VestingEntries.VestingEntry[] memory vestingEntries);
 }
